move frame-continuation and -header into own headers
